+++ lib/src/job/event/parallel.dart
@@ -2,10 +2,9 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
-// Dart imports:
-import 'dart:async';
-
 // Project imports:
+import 'package:batch/src/job/config/retry_configuration.dart';
+import 'package:batch/src/job/config/skip_configuration.dart';
 import 'package:batch/src/job/context/execution_context.dart';
 import 'package:batch/src/job/event/event.dart';
 import 'package:batch/src/job/task/parallel_task.dart';
@@ -15,16 +14,16 @@ class Parallel extends Event<Parallel> {
   Parallel({
     required String name,
     required List<ParallelTask> tasks,
-    FutureOr<bool> Function()? precondition,
     Function(ExecutionContext context)? onStarted,
     Function(ExecutionContext context)? onSucceeded,
     Function(ExecutionContext context, dynamic error, StackTrace stackTrace)?
         onError,
     Function(ExecutionContext context)? onCompleted,
+    SkipConfiguration? skipConfig,
+    RetryConfiguration? retryConfig,
   })  : _tasks = tasks,
         super(
           name: name,
-          precondition: precondition,
           onStarted: onStarted,
           onSucceeded: onSucceeded,
           onError: onError,